+++ .env.default
@@ -1,0 +1,10 @@
+#Port number for the server
+PORT=
+# Authentication for Postgress
+DATABASE=
+DATABASE_URL=
+DATABASE_USER=
+DATABASE_PASSWORD=
+# User Authentication
+PRIVATE_KEY=
+PUBLIC_KEY=
