+++ README.md
@@ -1,1 +1,42 @@
-# server
+# Server
+
+> Talent Path Pipeline Back-end
+
+## Getting Started
+
+### Setting up for local development
+
+1. Clone the repo to your machine
+
+2. Run the following command ```npm i``` to install all packages
+
+3. Create a .env file on the root directory with the information from .env.default file
+
+4. Have [Postgress running](https://www.postgresql.org/docs/current/server-start.html) in the background.
+
+5. Run  ```npm start``` to run the server
+
+6. Sequelize will generate all the tables for Postgress before the server begins listening for requests.
+
+   
+
+### Setting up for deployment
+
+**Coming Soon!**
+
+
+
+## Basic Structure
+
+The back-end is structure like the following:
+
+```text
+// Coming Soon!
+	
+```
+
+## Usage
+
+The following routes are available to be used:
+
+**Coming Soon!**
