+++ validations/users.validation.js
@@ -1,0 +1,8 @@
+function emailValidation(email) {
+  // This is a placeholder
+  return true;
+}
+function passwordValidation(email) {
+  // This is a placeholder
+  return true;
+}
